requirements.txt
aniso8601==9.0.1
autopep8==2.0.2
beautifulsoup4==4.12.2
bidict==0.22.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
dlib==19.24.1
dnspython==2.3.0
ExifRead==3.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.72
piexif==1.1.3
Pillow==9.5.0
pycodestyle==2.10.0
PyJWT==2.7.0
pymongo==4.3.3
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
tomli==2.0.1
ua-parser==0.16.1
urllib3==1.26.15
user-agents==2.2.0
Werkzeug==2.2.3
